+++ src/Illuminate/Database/Connectors/MySqlConnector.php
@@ -6,7 +6,7 @@ class MySqlConnector extends Connector implements ConnectorInterface {
 	 * Establish a database connection.
 	 *
 	 * @param  array  $config
-	 * @return PDO
+	 * @return \PDO
 	 */
 	public function connect(array $config)
 	{
