+++ src/Illuminate/Database/Connectors/SQLiteConnector.php
@@ -6,7 +6,7 @@ class SQLiteConnector extends Connector implements ConnectorInterface {
 	 * Establish a database connection.
 	 *
 	 * @param  array  $config
-	 * @return PDO
+	 * @return \PDO
 	 *
 	 * @throws \InvalidArgumentException
 	 */
