+++ src/Illuminate/Database/Eloquent/Builder.php
@@ -235,7 +235,7 @@ class Builder {
 	/**
 	 * Get a paginator for a grouped statement.
 	 *
-	 * @param  \Illuminate\Pagination\Environment  $paginator
+	 * @param  \Illuminate\Pagination\Factory  $paginator
 	 * @param  int    $perPage
 	 * @param  array  $columns
 	 * @return \Illuminate\Pagination\Paginator
@@ -250,7 +250,7 @@ class Builder {
 	/**
 	 * Get a paginator for an ungrouped statement.
 	 *
-	 * @param  \Illuminate\Pagination\Environment  $paginator
+	 * @param  \Illuminate\Pagination\Factory  $paginator
 	 * @param  int    $perPage
 	 * @param  array  $columns
 	 * @return \Illuminate\Pagination\Paginator
