+++ src/Illuminate/Exception/WhoopsDisplayer.php
@@ -38,6 +38,7 @@ class WhoopsDisplayer implements ExceptionDisplayerInterface {
 	 * Display the given exception to the user.
 	 *
 	 * @param  \Exception  $exception
+	 * @return \Symfony\Component\HttpFoundation\Response
 	 */
 	public function display(Exception $exception)
 	{
