+++ src/Illuminate/Pagination/PaginationServiceProvider.php
@@ -20,7 +20,7 @@ class PaginationServiceProvider extends ServiceProvider {
 	{
 		$this->app->bindShared('paginator', function($app)
 		{
-			$paginator = new Environment($app['request'], $app['view'], $app['translator']);
+			$paginator = new Factory($app['request'], $app['view'], $app['translator']);
 
 			$paginator->setViewName($app['config']['view.pagination']);
