+++ src/Illuminate/Pagination/Paginator.php
@@ -11,11 +11,11 @@ use Illuminate\Support\Contracts\ArrayableInterface;
 class Paginator implements ArrayableInterface, ArrayAccess, Countable, IteratorAggregate, JsonableInterface {
 
 	/**
-	 * The pagination environment.
+	 * The pagination factory.
 	 *
-	 * @var \Illuminate\Pagination\Environment
+	 * @var \Illuminate\Pagination\Factory
 	 */
-	protected $env;
+	protected $factory;
 
 	/**
 	 * The items being paginated.
@@ -83,16 +83,16 @@ class Paginator implements ArrayableInterface, ArrayAccess, Countable, IteratorA
 	/**
 	 * Create a new Paginator instance.
 	 *
-	 * @param  \Illuminate\Pagination\Environment  $env
+	 * @param  \Illuminate\Pagination\Factory  $factory
 	 * @param  array  $items
 	 * @param  int    $total
 	 * @param  int    $perPage
 	 * @return void
 	 */
-	public function __construct(Environment $env, array $items, $total, $perPage)
+	public function __construct(Factory $factory, array $items, $total, $perPage)
 	{
-		$this->env = $env;
 		$this->items = $items;
+		$this->factory = $factory;
 		$this->total = (int) $total;
 		$this->perPage = (int) $perPage;
 	}
@@ -143,7 +143,7 @@ class Paginator implements ArrayableInterface, ArrayAccess, Countable, IteratorA
 	 */
 	protected function calculateCurrentPage($lastPage)
 	{
-		$page = $this->env->getCurrentPage();
+		$page = $this->factory->getCurrentPage();
 
 		// The page number will get validated and adjusted if it either less than one
 		// or greater than the last page available based on the count of the given
@@ -175,7 +175,7 @@ class Paginator implements ArrayableInterface, ArrayAccess, Countable, IteratorA
 	 */
 	public function links($view = null)
 	{
-		return $this->env->getPaginationView($this, $view);
+		return $this->factory->getPaginationView($this, $view);
 	}
 
 	/**
@@ -187,7 +187,7 @@ class Paginator implements ArrayableInterface, ArrayAccess, Countable, IteratorA
 	public function getUrl($page)
 	{
 		$parameters = array(
-			$this->env->getPageName() => $page,
+			$this->factory->getPageName() => $page,
 		);
 
 		// If we have any extra query string key / value pairs that need to be added
@@ -200,7 +200,7 @@ class Paginator implements ArrayableInterface, ArrayAccess, Countable, IteratorA
 
 		$fragment = $this->buildFragment();
 
-		return $this->env->getCurrentUrl().'?'.http_build_query($parameters, null, '&').$fragment;
+		return $this->factory->getCurrentUrl().'?'.http_build_query($parameters, null, '&').$fragment;
 	}
 
 	/**
@@ -377,17 +377,17 @@ class Paginator implements ArrayableInterface, ArrayAccess, Countable, IteratorA
 	*/
 	public function setBaseUrl($baseUrl)
 	{
-		$this->env->setBaseUrl($baseUrl);
+		$this->factory->setBaseUrl($baseUrl);
 	}
 
 	/**
-	 * Get the pagination environment.
+	 * Get the pagination factory.
 	 *
-	 * @return \Illuminate\Pagination\Environment
+	 * @return \Illuminate\Pagination\Factory
 	 */
-	public function getEnvironment()
+	public function getFactory()
 	{
-		return $this->env;
+		return $this->factory;
 	}
 
 	/**
