+++ src/Illuminate/Queue/IronQueue.php
@@ -101,7 +101,7 @@ class IronQueue extends Queue implements QueueInterface {
 	 *
 	 * @param  \DateTime|int  $delay
 	 * @param  string  $job
-	 * @param  mixed  data
+	 * @param  mixed  $data
 	 * @param  string  $queue
 	 * @return mixed
 	 */
