+++ src/Illuminate/Support/Facades/Paginator.php
@@ -1,7 +1,7 @@
 <?php namespace Illuminate\Support\Facades;
 
 /**
- * @see \Illuminate\Pagination\Environment
+ * @see \Illuminate\Pagination\Factory
  */
 class Paginator extends Facade {
