+++ src/Illuminate/Support/Facades/Response.php
@@ -65,7 +65,7 @@ class Response {
 	/**
 	 * Return a new streamed response from the application.
 	 *
-	 * @param  Closure  $callback
+	 * @param  \Closure  $callback
 	 * @param  int      $status
 	 * @param  array    $headers
 	 * @return \Symfony\Component\HttpFoundation\StreamedResponse
@@ -78,7 +78,7 @@ class Response {
 	/**
 	 * Create a new file download response.
 	 *
-	 * @param  SplFileInfo|string  $file
+	 * @param  \SplFileInfo|string  $file
 	 * @param  string  $name
 	 * @param  array   $headers
 	 * @return \Symfony\Component\HttpFoundation\BinaryFileResponse
