+++ src/Illuminate/View/Engines/PhpEngine.php
@@ -47,7 +47,7 @@ class PhpEngine implements EngineInterface {
 	/**
 	 * Handle a view exception.
 	 *
-	 * @param  Exception  $e
+	 * @param  \Exception  $e
 	 * @return void
 	 *
 	 * @throws $e
